requirements.txt
alpaca_trade_api==3.2.0
python-dotenv==1.1.1
